Fixing unwanted click events on map drag
The JS click event is fired if the element which the click event is
bound to receives a mousedown followed by a mouseup. This happens
regardless of the time between the mouse down and up, or of any mouse
motion—so long as the cursor was within the element's boundaries at the
time of the down and the up, it will fire a click.

When dragging the map, it sometimes caused unwanted click events while
dragging the map, particularly when zoomed in (causing the clickable
elements to be larger, and raising the chances that the cursor would
still be inside them when the mouse button was released).

The solution listens separately for mousedown and mouseup, and takes
action only if there's been no motion between mousedown and mouseup.